requirements.txt
Django==3.1.4
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
pytz==2019.3
psycopg2==2.8.6